Add need for confirmation before patching. [#68]

+++ Functions/Patching.py
@@ -76,8 +76,12 @@ def patch(hexdump, address, bytes_to_write, architecture, file_name, colors):
             return 1
         bytes_list[bytes_pointer]=int(bytes_to_write[i+i:i+i+2],16)
         bytes_pointer+=1
-    file_o = open(file_name, 'wb')
-    file_o.write(bytes(bytes_list))
-    print("Patching succed.")
-    print("Old bytes: "+str(''.join(old_bytes)))
-    print("New bytes: "+bytes_to_write)
+    userConfirmation = input("You WILL change the file, do you want to proceed? (Y/n): ")
+    if userConfirmation=='' or userConfirmation=="Y" or userConfirmation=="y" or userConfirmation=="Yes" or userConfirmation=="yes":
+        file_o = open(file_name, 'wb')
+        file_o.write(bytes(bytes_list))
+        print("Patching succed.")
+        print("Old bytes: "+str(''.join(old_bytes)))
+        print("New bytes: "+bytes_to_write)
+    else:
+        print("Canceled")
